+++ src/main/java/kr/shihyeon/imagicthud/gui/screen/modmenu/ModMenuIntegration.java
@@ -1,8 +1,8 @@
-package kr.shihyeon.imagicthud.gui.config.modmenu;
+package kr.shihyeon.imagicthud.gui.screen.modmenu;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import kr.shihyeon.imagicthud.gui.config.yacl3.YaclConfigScreenFactoryManager;
+import kr.shihyeon.imagicthud.gui.screen.config.yacl3.YaclConfigScreenFactoryManager;
 
 public class ModMenuIntegration implements ModMenuApi {
