+++ src/main/java/kr/shihyeon/imagicthud/mixin/gui/sodium/SodiumOptionsGUIMixin.java
@@ -1,6 +1,6 @@
 package kr.shihyeon.imagicthud.mixin.gui.sodium;
 
-import kr.shihyeon.imagicthud.gui.config.sodium.SodiumOptionPages;
+import kr.shihyeon.imagicthud.gui.screen.config.sodium.SodiumOptionPages;
 import net.caffeinemc.mods.sodium.client.gui.SodiumOptionsGUI;
 import net.caffeinemc.mods.sodium.client.gui.options.OptionPage;
 import org.spongepowered.asm.mixin.Final;
